+++ test/vagrant/action/vm/share_folders_test.rb
@@ -114,8 +114,8 @@ class ShareFoldersVMActionTest < Test::Unit::TestCase
   context "mounting the shared folders" do
     setup do
       @folders = stub_shared_folders(<<-sf)
-        config.vm.share_folder("foo", "fooguest", "foohost")
-        config.vm.share_folder("bar", "barguest", "barhost")
+        config.vm.share_folder("foo", "fooguest", "foohost", :owner => "yo", :group => "fo")
+        config.vm.share_folder("bar", "barguest", "barhost", :owner => "foo", :group => "bar")
         config.vm.share_folder("foo_no_mount", nil, "foohost2")
       sf
       @ssh = mock("ssh")
@@ -127,9 +127,9 @@ class ShareFoldersVMActionTest < Test::Unit::TestCase
       mount_seq = sequence("mount_seq")
       @folders.each do |name, data|
         if data[:guestpath]
-          @vm.system.expects(:mount_shared_folder).with(@ssh, name, data[:guestpath]).in_sequence(mount_seq)
+          @vm.system.expects(:mount_shared_folder).with(@ssh, name, data[:guestpath], data[:owner], data[:group]).in_sequence(mount_seq)
         else
-          @vm.system.expects(:mount_shared_folder).with(@ssh, name, anything).never
+          @vm.system.expects(:mount_shared_folder).with(@ssh, name, anything, anything, anything).never
         end
       end